Refactor pauser privilege level test into separate file
