+++ runtime/tests/runtime_integration_tests/main.rs
@@ -11,6 +11,7 @@ mod test_info;
 mod test_invoke_dpe;
 mod test_mailbox;
 mod test_panic_missing;
+mod test_pauser_privilege_levels;
 mod test_populate_idev;
 mod test_stash_measurement;
 mod test_tagging;
